Tried to implement write/read methods, also refactored old code

+++ FileSystem/file_system/entities/DirectoryEntry.cpp
@@ -13,7 +13,7 @@ DirectoryEntry::~DirectoryEntry() {
 }
 
 void DirectoryEntry::parse(const char *bytes) {
-    for (size_t i = 0; i < 4; i++) {
+    for (int i = 0; i < 4; i++) {
         file_name[i] = bytes[i];
     }
     descriptor_index = Utils::bytesToInt32(bytes + 4);
@@ -24,7 +24,7 @@ bool DirectoryEntry::isFree() const {
 }
 
 void DirectoryEntry::copyFileName(char *buffer) const {
-    for (size_t i = 0; i < 5; i++) {
+    for (int i = 0; i < 5; i++) {
         buffer[i] = file_name[i];
     }
 }
@@ -34,7 +34,7 @@ int DirectoryEntry::getDescriptorIndex() const {
 }
 
 void DirectoryEntry::setFileName(const char *new_file_name) {
-    for (size_t i = 0; i < 4; i++) {
+    for (int i = 0; i < 4; i++) {
         file_name[i] = new_file_name[i];
         if (new_file_name[i] == '\0') {
             break;
@@ -47,12 +47,12 @@ void DirectoryEntry::setDescriptorIndex(int new_descriptor_index) {
 }
 
 void DirectoryEntry::copyBytes(char *buffer) const {
-    for (size_t i = 0; i < 4; i++) {
+    for (int i = 0; i < 4; i++) {
         buffer[i] = file_name[i];
     }
 
     char *descriptor_index_bytes = Utils::intToBytes(descriptor_index);
-    for (size_t i = 0; i < 4; i++) {
+    for (int i = 0; i < 4; i++) {
         buffer[4 + i] = descriptor_index_bytes[i];
     }
     delete[] descriptor_index_bytes;
